http://jira.gapso.com.br/browse/TRIEDA-1376

+++ code/solver/Solver Trieda/Calendario.cpp
@@ -159,7 +159,7 @@ bool Calendario::intervaloEntreHorarios( HorarioAula *hi, HorarioAula *hf )
 }
 
  
-int Calendario::getTempoTotal( int dia )
+double Calendario::getTempoTotal( int dia )
 {
 	return ( tempo_aula * this->getNroDeHorariosAula(dia) ); 
 }
